pyproject.toml
[project]
name = "rag-citation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.45.2",
    "markitdown>=0.0.1a3",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
]
